refactor: WinnerLottery -> LotteryMatcher 로 이동함에 따른 테스트 코드 수정
