+++ src/test/kotlin/lottery/domain/LotteryMatcherTest.kt
@@ -1,0 +1,30 @@
+package lottery.domain
+
+import org.assertj.core.api.Assertions
+import org.junit.jupiter.api.Test
+
+class LotteryMatcherTest {
+    @Test
+    fun `로또번호들 중에 당첨번호와 일치하는 개수들을 전달한다`() {
+        val lotteries = listOf(
+            createLottery(listOf(1, 3, 5, 2, 9, 40)),
+            createLottery(listOf(34, 24, 1, 15, 3, 5)),
+            createLottery(listOf(3, 1, 5, 15, 24, 40)),
+            createLottery(listOf(1, 3, 5, 40, 24, 15))
+        )
+
+        val winnerLottery = WinnerLottery(listOf(1, 3, 5, 24, 15, 40))
+
+        val lotteryMatcher = LotteryMatcher(winnerLottery, lotteries)
+
+        val matchCounts = lotteryMatcher.match()
+
+        org.junit.jupiter.api.assertAll(
+            { Assertions.assertThat(matchCounts[4]).isEqualTo(1) },
+            { Assertions.assertThat(matchCounts[5]).isEqualTo(1) },
+            { Assertions.assertThat(matchCounts[6]).isEqualTo(2) }
+        )
+    }
+
+    private fun createLottery(numbers: List<Int>) = Lottery(LotteryNumbers(numbers))
+}
